Extract example generation to TourExampleGeneration
