+++ apps/antalmanac/src/lib/tourExampleGeneration.ts
@@ -6,6 +6,56 @@ import {
     WebsocSectionMeeting,
     daysOfWeek,
 } from 'peterportal-api-next-types';
+import AppStore from '$stores/AppStore';
+
+export function addSampleClasses() {
+    if (AppStore.getAddedCourses().length > 0) return;
+
+    const sampleClasses: Array<ScheduleCourse> = [
+        {
+            courseTitle: 'Nice',
+            deptCode: 'GEN&SEX',
+            courseNumber: '69',
+            instructors: ['Your mother'],
+            meetings: [
+                {
+                    bldg: ['DBH'],
+                    days: 'MWF',
+                    startTime: {
+                        hour: 10,
+                        minute: 0,
+                    },
+                    endTime: {
+                        hour: 10,
+                        minute: 50,
+                    },
+                    timeIsTBA: false,
+                },
+            ],
+        },
+        {
+            meetings: [
+                {
+                    bldg: ['ELH 100'],
+                    days: 'TuTh',
+                    startTime: {
+                        hour: 9,
+                        minute: 30,
+                    },
+                    endTime: {
+                        hour: 10,
+                        minute: 50,
+                    },
+                    timeIsTBA: false,
+                },
+            ],
+        },
+    ].map(sampleClassFactory);
+
+    sampleClasses.forEach((sampleClass) => {
+        AppStore.addCourse(sampleClass);
+    });
+}
 
 export function randint(min: number, max: number): number {
     return Math.floor(Math.random() * (max - min + 1)) + min;
